🩹 MemberService interface 제거
