+++ src/main/java/knu/team1/be/boost/member/service/MemberService.java
@@ -3,12 +3,38 @@ package knu.team1.be.boost.member.service;
 import java.util.UUID;
 import knu.team1.be.boost.member.dto.MemberResponseDto;
 import knu.team1.be.boost.member.dto.MemberUpdateRequestDto;
+import knu.team1.be.boost.member.entity.Member;
+import knu.team1.be.boost.member.exception.MemberNotFoundException;
+import knu.team1.be.boost.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface MemberService {
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class MemberService {
 
-    MemberResponseDto getMember(UUID userId);
+    private final MemberRepository memberRepository;
 
-    MemberResponseDto updateMember(UUID userId, MemberUpdateRequestDto requestDto);
+    public MemberResponseDto getMember(UUID userId) {
+        Member member = memberRepository.findById(userId)
+            .orElseThrow(() -> new MemberNotFoundException(userId));
+        return MemberResponseDto.from(member);
+    }
 
-    void deleteMember(UUID userId);
+    @Transactional
+    public MemberResponseDto updateMember(UUID userId, MemberUpdateRequestDto requestDto) {
+        Member member = memberRepository.findById(userId)
+            .orElseThrow(() -> new MemberNotFoundException(userId));
+        member.updateMember(requestDto.name(), requestDto.avatar());
+        return MemberResponseDto.from(member);
+    }
+
+    @Transactional
+    public void deleteMember(UUID userId) {
+        Member member = memberRepository.findById(userId)
+            .orElseThrow(() -> new MemberNotFoundException(userId));
+        memberRepository.delete(member);
+    }
 }
